refactor comment create
