+++ server/services/comment/comment.service.js
@@ -1,6 +1,9 @@
+const mongoose = require("mongoose");
 const { User } = require("../../models/user");
 const { Comment } = require("../../models/comment");
+const { Post } = require("../../models/post");
 const { getPagination } = require("../../utils/pagination");
+const { NotFoundException } = require("../../utils/exceptions");
 
 module.exports = class CommentService {
   COMMENTS_PER_PAGE = 10;
@@ -22,6 +25,22 @@ module.exports = class CommentService {
 
     return { comments, pageInfo };
   }
+
+  async create({ authorId, postId, commentBody }) {
+    let comment;
+    const session = await mongoose.startSession();
+    session.startTransaction();
+    try {
+      comment = await createComment(authorId, postId, commentBody);
+      await session.commitTransaction();
+      session.endSession();
+      return comment;
+    } catch (error) {
+      await session.abortTransaction();
+      session.endSession();
+      throw error;
+    }
+  }
 };
 
 async function parseQuery({ body, author }) {
@@ -42,4 +61,33 @@ function setCommentBodyQuery(body, query) {
   if (body) {
     query.body = { $regex: body };
   }
+}
+
+async function createComment(authorId, postId, commentBody) {
+  const commentData = {
+    author: authorId,
+    post: postId,
+    score: 0,
+    body: commentBody,
+  };
+  const comment = await Comment.create(commentData);
+
+  await addCommentToPost(postId, comment);
+
+  return comment;
+}
+
+async function addCommentToPost(postId, comment) {
+  const post = await getPost(postId);
+  post.comments.push(comment);
+  await post.save();
+}
+
+async function getPost(id) {
+  const post = await Post.findById(id);
+  if (!post) {
+    const message = `Post ${id} does not exist`;
+    throw new NotFoundException(message);
+  }
+  return post;
 }
